refactor: all @AdminOnly handler methods are now in  AdminController.java
