+++ src/main/java/com/example/roommate/controller/RoomController.java
@@ -2,7 +2,6 @@ package com.example.roommate.controller;
 
 import com.example.roommate.annotations.AdminOnly;
 import com.example.roommate.application.services.AdminApplicationService;
-import com.example.roommate.exceptions.NotFoundRepositoryException;
 import com.example.roommate.values.domainValues.*;
 import com.example.roommate.exceptions.applicationService.NotFoundException;
 import com.example.roommate.interfaces.entities.IRoom;
@@ -130,24 +129,5 @@ public class RoomController {
     }
 
 
-    @AdminOnly
-    @PostMapping("/room/{roomID}/addItem/{itemName}")
-    public ModelAndView addItem(Model model, @PathVariable UUID roomID, @PathVariable String itemName) throws NotFoundRepositoryException {
-        bookingApplicationService.addItemToRoom(roomID, itemName);
-        return roomDetails(model, roomID);
-    }
 
-    @AdminOnly
-    @PostMapping("/room/{roomID}/createItem")
-    public ModelAndView createItem(Model model, @PathVariable UUID roomID, @RequestParam String newItem) throws NotFoundRepositoryException {
-        bookingApplicationService.createItem(newItem);
-        return addItem(model, roomID, newItem);
-    }
-
-    @AdminOnly
-    @PostMapping("/room/{roomID}/deleteItem/{itemName}")
-    public ModelAndView deleteItem(Model model, @PathVariable UUID roomID, @PathVariable String itemName) throws NotFoundRepositoryException {
-        bookingApplicationService.removeItemFromRoom(roomID, itemName);
-        return roomDetails(model, roomID);
-    }
 }
